move and cleanup SubProcCopyGlobalConfigPreInitFunc
