+++ returnn/torch/data/pipeline.py
@@ -319,13 +319,14 @@ def create_data_loader_from_batches(
         # (https://github.com/rwth-i6/returnn/issues/1495).
         loader_opts.setdefault("multiprocessing_context", "spawn_non_daemonic")
         if loader_opts.get("multiprocessing_context") == "spawn_non_daemonic":
+            from returnn.config import SubProcCopyGlobalConfigPreInitFunc
             from returnn.util.multi_proc_non_daemonic_spawn import NonDaemonicSpawnContext
 
             loader_opts["multiprocessing_context"] = NonDaemonicSpawnContext(
                 # It's important that this init function is called even before the unpickling of the dataset,
                 # as that unpickling might depend on the right context, e.g. having a global RETURNN config.
                 # See _DataLoaderWorkerPreInitFunc below, https://github.com/rwth-i6/returnn/issues/1495.
-                process_pre_init_func=_DataLoaderWorkerPreInitFunc()
+                process_pre_init_func=SubProcCopyGlobalConfigPreInitFunc()
             )
 
     return torch.utils.data.DataLoader(
@@ -341,31 +342,6 @@ def create_data_loader_from_batches(
         # User-defined
         **loader_opts,
     )
-
-
-class _DataLoaderWorkerPreInitFunc:
-    def __init__(self):
-        # Get the RETURNN global config here. Allow this to be optional (for use outside of RETURNN).
-        # We store it here such that pickling this worker init func will also pickle the config,
-        # so that we can reset it as global config inside the worker.
-        # Some RETURNN datasets might depend on the config.
-        # https://github.com/rwth-i6/returnn/issues/1495
-        # MultiProcDataset has a similar logic, see https://github.com/rwth-i6/returnn/issues/1384.
-        from returnn.config import get_global_config
-
-        self.global_config = get_global_config(raise_exception=False)
-
-    def __call__(self):
-        from returnn.util import better_exchook
-        from returnn import __old_mod_loader__
-
-        better_exchook.install()
-        __old_mod_loader__.disable_lazy_mod_loads()
-
-        if self.global_config:
-            from returnn.config import set_global_config
-
-            set_global_config(self.global_config)
 
 
 class _DataLoaderWorkerInitFunc:
